relocate random number generation codes
* moves missing random123 functions and our generator functors from base
  into r123ext
* improves the documentation of these codes
