+++ RandBLAS/base.hh
@@ -65,95 +65,6 @@ std::ostream &operator<<(
     return out;
 }
 
-
-/** Apply boxmuller transform to all elements of r. The number of elements of r
- * must be evenly divisible by 2.
- *
- * @tparam RNG a random123 CBRNG type
- * @tparam T the desired return type. The default return type is dictated by
- *           the RNG's counter element type. float for 32 bit counter elements
- *           and double for 64.
- *
- * @param[in] ri a sequence of n random values generated using random123 CBRNG
- *               type RNG. The transform is applied pair wise.
- *
- * @returns n transformed floating point values.
- */
-template <typename RNG, typename T = typename std::conditional
-    <sizeof(typename RNG::ctr_type::value_type) == sizeof(uint32_t), float, double>::type>
-auto boxmulall(
-    typename RNG::ctr_type const& ri
-) {
-    std::array<T, RNG::ctr_type::static_size> ro;
-    int nit = RNG::ctr_type::static_size / 2;
-    for (int i = 0; i < nit; ++i)
-    {
-        auto [v0, v1] = r123::boxmuller(ri[2*i], ri[2*i + 1]);
-        ro[2*i    ] = v0;
-        ro[2*i + 1] = v1;
-    }
-    return ro;
-}
-
-/// Generate a sequence of random values and apply a Box-Muller transform.
-struct boxmul
-{
-    /** Generate a sequence of random values and apply a Box-Muller transform.
-     *
-     * @tparam RNG a random123 CBRNG type
-     * @tparam T the desired return type. The default return type is dictated by
-     *           the RNG's counter element type. float for 32 bit counter elements
-     *           and double for 64.
-     *
-     * @param[in] a random123 CBRNG instance used to generate the sequence
-     * @param[in] the CBRNG counter of n elements
-     * @param[in] the CBRNG key
-     *
-     * @returns the generated and transformed sequence of n floating point
-     *          values. The return type is dictated by the RNG's counter
-     *          element type. float for 32 bit counter elements and double for
-     *          64.
-     */
-    template <typename RNG>
-    static
-    auto generate(
-        RNG &rng,
-        typename RNG::ctr_type const& c,
-        typename RNG::key_type const& k
-    ) {
-        return boxmulall<RNG>(rng(c,k));
-    }
-};
-
-/// Generate a sequence of random values and transform to -1.0 to 1.0.
-struct uneg11
-{
-    /** Generate a sequence of random values and transform to -1.0 to 1.0.
-     *
-     * @tparam RNG a random123 CBRNG type
-     * @tparam T the desired return type. The default return type is dictated by
-     *           the RNG's counter element type. float for 32 bit counter elements
-     *           and double for 64.
-     *
-     * @param[in] a random123 CBRNG instance used to generate the sequence
-     * @param[in] the CBRNG counter of n elements
-     * @param[in] the CBRNG key
-     *
-     * @returns the generated and transformed sequence of n floating point
-     *          values. The return type is dictated by the counter element
-     *          type. float for 32 bit counter elements and double for 64.
-     */
-    template <typename RNG, typename T = typename std::conditional
-        <sizeof(typename RNG::ctr_type::value_type) == sizeof(uint32_t), float, double>::type>
-    static
-    auto generate(
-        RNG &rng,
-        typename RNG::ctr_type const& c,
-        typename RNG::key_type const& k
-    ) {
-        return r123::uneg11all<T>(rng(c,k));
-    }
-};
-}; // end namespace RandBLAS::base
+} // end namespace RandBLAS::base
 
 #endif
